fixes bug where user's friends were not being retrieved from strava
